minor updates
build time print, game / network check, removed rfid references in mqtt network class, updated fx450.h to support rev 1, updated LED references

+++ game.h
@@ -124,11 +124,12 @@ class Game
 	uint8_t FX60_1_RSTATE_OLD = 0;
 #endif
 
-	void EEPROMReadString(byte pos, byte len, char* data);
-	void EEPROMWriteString(byte pos, char* data);
+	void EEPROMReadString(size_t pos, byte len, char* data);
+	void EEPROMWriteString(size_t pos, char* data);
 
   private:
 	  
 };
 
 #endif
+
